# [skrifa] more scaler perf