Follow up to #1171 for even faster outline loading:

* Some targeted inlining, specifically `<i8 as Scalar>::from_raw` was showing up in the profile. This is a big win
* Use a borrow of the largish `glyf::Outlines` type when constructing a scaler to save the copy.. removes some `_platform_memcpy`
* Add bucketed stack allocations to avoid zeroing excessive memory.. significantly reduces cost of `_platform_memset` in the profile

These changes reduce average time for loading all Noto Sans Regular glyphs from 295ms (previous PR) to 254ms (206ms with LTO). This is down from our starting point of 337ms.